refactor(nicknamer): move `commands` module to `user`
Relocated the `commands` module functionality to a new `user` module for better clarity and alignment with the refactored architecture. Adjusted imports and references across the codebase to maintain consistency.
