+++ nicknamer/src/nicknamer/user.rs
